tmp: now need to make the cards service work
Multiple options:
- duplicate the service in each window, so that we keep everything synchronous,
but have higher mem consumption
- use a single reference service (in MainWindow or in the main process), and
use ipc to communicate back and forth. All calls become async, which means a
lot of rewrites throughout the app
This will also require the base service in MainWindow to be able to listen to
electron calls, so maybe use a subclass that has this logic?

Same question for i18n.